+++ app/lib/features/my_bookings/presentation/manager/booking_detail_bloc/booking_detail_event.dart
@@ -1,0 +1,16 @@
+part of 'booking_detail_bloc.dart';
+
+sealed class BookingDetailEvent extends Equatable {
+  const BookingDetailEvent();
+
+  @override
+  List<Object?> get props => [];
+}
+
+final class LoadBookingDetailEvent extends BookingDetailEvent {
+  const LoadBookingDetailEvent();
+}
+
+final class CancelBookingEvent extends BookingDetailEvent {
+  const CancelBookingEvent();
+}